Modularise pricing and add `Pricing` trait
